==> src/1-cleanning_artist.ipynb <==
import pandas as pd
import numpy as np

artists_top1 = pd.read_csv(r"/mnt/c/Users/Palmira/Desktop/Ironhack/Semana4/project_ETL/data/10000-MTV-Music-Artists-page-1.csv",encoding = "ISO-8859-1")
artists_top2 = pd.read_csv(r"/mnt/c/Users/Palmira/Desktop/Ironhack/Semana4/project_ETL/data/10000-MTV-Music-Artists-page-2.csv",encoding = "ISO-8859-1")
artists_top3 = pd.read_csv(r"/mnt/c/Users/Palmira/Desktop/Ironhack/Semana4/project_ETL/data/10000-MTV-Music-Artists-page-3.csv",encoding = "ISO-8859-1")
artists_top4 = pd.read_csv(r"/mnt/c/Users/Palmira/Desktop/Ironhack/Semana4/project_ETL/data/10000-MTV-Music-Artists-page-4.csv",encoding = "ISO-8859-1")
artists= [artists_top1,artists_top2,artists_top3,artists_top4]

result = pd.concat(artists)

result.rename(columns={'name': 'artist'}, inplace=True)
result.sample(7)

result.index.set_names(['ranking_MTV'], inplace=True)

result.dropna(subset=["artist"], inplace = True)
result.shape

result.drop_duplicates(subset=["artist"], inplace=True)
result.shape

result.to_csv(r'/mnt/c/Users/Palmira/Desktop/Ironhack/Semana4/project_ETL/src/Output_data/artits_MTV.csv', index=False)
