==> src/2-scrapping_top_streamed.ipynb <==
import requests as req

from bs4 import BeautifulSoup as bs

import pandas as pd

#most streamed spotify artists
url = 'https://chartmasters.org/most-streamed-artists-ever-on-spotify/'

html = req.get(url).content
soup = bs(html, 'html.parser')

artist_spoti = pd.read_html(html)[0]
artist_spoti.sample()

artist_spoti.rename(columns={"#":"streaming_rank",'Artist Name': 'artist', "Lead Streams":"lead_streams", "Featured Streams":"featured_streams", "Last Update":"last_update"}, inplace=True)

artist_spoti.dropna(subset=["artist"], inplace = True)
artist_spoti.shape

artist_spoti.drop_duplicates(subset=["artist"], inplace=True)

artist_spoti

artist_spoti.to_csv(r'/mnt/c/Users/Palmira/Desktop/Ironhack/Semana4/project_ETL/src/Output_data/artits_streamed_spoti.csv', index=False)
