==> src/3-spotipy_loop.ipynb <==
import json
import requests as req
import pandas as pd
import os
import logging
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
from spotipy.oauth2 import SpotifyOAuth
from webdriver_manager.chrome import ChromeDriverManager
import requests as req
from bs4 import BeautifulSoup as bs
import pandas as pd
import selenium
import json 
from selenium import webdriver
import os
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.common.by import By
import time
import sys

with open('secret2.txt') as f:
    secret2 = f.readlines()

sp = spotipy.Spotify(auth_manager=SpotifyClientCredentials(client_id="d663de0869444f16aa6d03d9c8ae4fe3",
                                                           client_secret=secret2))

artists_spoti = pd.read_csv(r'/mnt/c/Users/Palmira/Desktop/Ironhack/Semana4/project_ETL/src/Output_data/artits_streamed_spoti.csv',encoding = "ISO-8859-1")

art1 = artists_spoti['artist'][0]
spoti_info = sp.search(q='Drake',limit=10, type='artist')
df = pd.DataFrame(spoti_info)
df.explode('artists').reset_index(drop=False)
df= pd.json_normalize(df.iloc[1].artists)
df1 = df[:1]
df1

%%time
lst = []
for i in artists_spoti['artist']:
    time.sleep(0.3)
    spoti_info = sp.search(q=i,limit=10, type='artist') #getting info from Spoti
    
    df = pd.DataFrame(spoti_info)  #Dataframe with all info  
    df.explode('artists').reset_index(drop=False)   #explode dataframe to see list of dataframes
    df2 = pd.json_normalize(df.iloc[1].artists) #choosing the column for the dataframe.
    
    artist2 = df2[:1]     #getting first artist on the search.
    
    df1 = pd.concat([df1,artist2], ignore_index=True)                #concatenate/Append artists dataframes.
        
df1.sample(3)

df1.head(3)

df1.drop(axis=0, index=None, columns = ["genres","images","href", "followers.href", "external_urls.spotify"], inplace=True)


df1.to_csv(r'/mnt/c/Users/Palmira/Desktop/Ironhack/Semana4/project_ETL/src/Output_data/spoti_info.csv')

Concatenando por medio de listas:

lst = []
for i in artists_spoti['artist']:
    time.sleep(0.5)
    spoti_info = sp.search(q=i,limit=10, type='artist') #getting info from Spoti
    
    df = pd.DataFrame(spoti_info)  #Dataframe with all info  
    df.explode('artists').reset_index(drop=False)   #explode dataframe to see list of dataframes
    df2 = pd.json_normalize(df.iloc[1].artists) #choosing the column for the dataframe.
    
    artist2 = df2[:1]     #getting first artist on the search.

    dic = {'name': artist2['name'][0],
           'genres': artist2['genres'][0],
          'id': artist2['id'][0],
          'images':artist2['images'][0],
          'type': artist2['type'][0],
           'uri': artist2['uri'][0],
           'followers.total':artist2['followers.total'][0]}
    
    lst.append(dic)   #concatenate/Append artists dataframes.
    

pd.DataFrame(lst)
