==> src/3.a-spotify-api.ipynb <==
## For checking  spotify go to spotipyloop.ipynb

import json
import requests as req
import pandas as pd
import os
import logging
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
from spotipy.oauth2 import SpotifyOAuth

from webdriver_manager.chrome import ChromeDriverManager
import requests as req
from bs4 import BeautifulSoup as bs
import pandas as pd
import selenium
import json 
from selenium import webdriver
import os
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.common.by import By
import time
import sys


with open('secret2.txt') as f:
    secret2 = f.readlines()

sp = spotipy.Spotify(auth_manager=SpotifyClientCredentials(client_id="d663de0869444f16aa6d03d9c8ae4fe3",
                                                           client_secret=secret2))



artists_spoti = pd.read_csv(r'/mnt/c/Users/Palmira/Desktop/Ironhack/Semana4/project_ETL/src/Output_data/artits_streamed_spoti.csv',encoding = "ISO-8859-1")

artists_spoti

art1 = artists_spoti['artist'][0]

spoti_info = sp.search(q=art1,limit=10, type='artist')

df = pd.DataFrame(spoti_info)


df.explode('artists').reset_index(drop=False)


df= pd.json_normalize(df.iloc[1].artists)

df.drop(axis=0, index=None, columns = ["href", "images", "external_urls.spotify"], inplace=True)

artist = df[:1]
artist

artist.rename(columns={"id":"artist_id"},inplace=True)

artist

results = sp.search(q='Drake')
results2={}


for idx, track in enumerate(results['tracks']['items']):
    print(idx, track['name'], track['popularity'])
    results2={'index' : idx,'Artist': track['name'],'popularity': track['popularity']}

results2
 
case_list={}
for idx, track in enumerate(results['tracks']['items']):
    case ={'index' : idx,'Artist': track['name'],'popularity': track['popularity']}
case  

adele = sp.search(q='Drake',limit=10, type='artist')
adele2= pd.json_normalize(adele)
#adele2.explode("artists.items")
adele2


artist_uri = artist['uri']

album = sp.artist_albums(artist_uri,limit=10, album_type='album')
df_alb = pd.DataFrame(album)
df_alb
df_alb= pd.json_normalize(df.iloc[1])
#df.explode('album').reset_index(drop=False)
#df.explode('artists').reset_index(drop=False)

artist_id='7MPGCB854Qo4alYMOPkBka'

artist = sp.artist('https://api.spotify.com/v1/artists/'+artist_id)
print(artist)


## Traernos los track id´s y pasarles un bucle en el que nos diera la informacion de spotify que queremos para todos. 

import pandas as pd
import mysql.connector as conn
from sqlalchemy import create_engine

db=conn.connect(host='localhost',
                     user='root',
                     passwd='admin',
                     database='Spotify')

cursor=db.cursor()

str_conn='mysql+pymysql://root:admin@localhost:3306/Spotify'

motor=create_engine(str_conn)  # motor de conexion

## NO SE PUEDEN SACAR LAS REPRODUCCIONES DE CANCIONES.

Aquí quiero hacer una búsqueda en selenium que me de las reproducciones de las canciones 

'https://open.spotify.com/search/track:%22Something%22%20artist:%22The%20Beatles%22'

### TO USE IN SECRET MODE:
with open('secret.txt') as f:
    secret = f.readline()
print (secret)

#sp = spotipy.Spotify(auth_manager=SpotifyOAuth(client_id="d663de0869444f16aa6d03d9c8ae4fe3",client_secret="47d3341ba76c4cf9a666c575b1c77ff9"))
