==> src/5-sql_queries.ipynb <==
import pandas as pd

artists_numbers = pd.read_csv("Output_data/artists_streamed_spoti.csv",encoding = "ISO-8859-1")
artists_numbers.drop_duplicates(inplace=True)
artists_numbers.drop_duplicates(subset='artist',inplace=True)

artists_numbers.rename(columns={"Tracks":"tracks",'1b+': '1b', "100m+":"100m", "10m+":"10m", "1m+":"1m"}, inplace=True)

artists_idSpotify = pd.read_csv("Output_data/spoti_info.csv",encoding = "ISO-8859-1")


artists_idSpotify.rename(columns={"followers.total":"followers_total"},inplace=True)

artists_idSpotify.drop_duplicates(inplace=True)
artists_idSpotify.duplicated().sum()

past_concerts = pd.read_csv("Output_data/past_concerts.csv",encoding = "ISO-8859-1")

upcoming_concerts = pd.read_csv("Output_data/upcoming_concerts.csv",encoding = "ISO-8859-1")

from sqlalchemy import create_engine
import pandas as pd

df = pd.read_csv('Output_data/artits_streamed_spoti.csv')
# Optional, set your indexes to get Primary Keys


engine = create_engine('mysql://user:pass@host/db', echo=False)

df.to_sql(table_name, dwh_engine, index=False)

import mysql.connector as conn

crea_db=conn.connect(host='localhost',
                     user='root',
                     passwd='admin',
                     database='Spotify')

cursor=crea_db.cursor()

#cursor.execute('drop database if exists spotify')
#cursor.execute('create database spotify')

cursor.execute('drop table if exists rank_streamed_artists')


tabla='''
            create table rank_streamed_artists(
                streaming_rank int,
                artist varchar(50),
                lead_streams varchar(20),
                featured_streams varchar(20),
                tracks int,
                1b int,
                100m int,
                10m int,
                1m int,
                last_update varchar(50)
            );
'''


cursor.execute(tabla)

for i in range(len(artists_numbers)):
    insert_query='insert into {} ({}) values {};'\
                 .format('rank_streamed_artists', 
                         ','.join(artists_numbers.columns), 
                         tuple(artists_numbers.iloc[i].values))
    
    cursor.execute(insert_query)
crea_db.commit()

query_key_stream='''
ALTER TABLE `Spotify`.`rank_streamed_artists` 
ADD PRIMARY KEY (`streaming_rank`);'''
cursor.execute(query_key_stream)

cursor.execute('drop table if exists past_concerts')


tabla_past_events='''
            create table past_concerts(
                int_artist int,
                artist varchar(50),
                date varchar(30),
                location varchar(300)
            );
'''

cursor.execute(tabla_past_events)

for i in range(len(past_concerts)):
    insert_query2='insert into {} ({}) values {};'\
                 .format('past_concerts', 
                         ','.join(past_concerts.columns), 
                         tuple(past_concerts.iloc[i].values))
    
    cursor.execute(insert_query2)
crea_db.commit()

query_key_past = '''
ALTER TABLE `Spotify`.`past_concerts` 
ADD PRIMARY KEY (`int_artist`);
'''
cursor.execute(query_key_past)

cursor.execute('drop table if exists upcoming_concerts')


tabla_upcoming_events='''
            create table upcoming_concerts(
                int_artist int,
                artist varchar(50),
                date varchar(30),
                location varchar(300)
            );
'''


cursor.execute(tabla_upcoming_events)

for i in range(len(upcoming_concerts)):
    insert_query2='insert into {} ({}) values {};'\
                 .format('upcoming_concerts', 
                         ','.join(upcoming_concerts.columns), 
                         tuple(upcoming_concerts.iloc[i].values))
    
    cursor.execute(insert_query2)
crea_db.commit()

query_key = '''
ALTER TABLE `Spotify`.`upcoming_concerts` 
ADD PRIMARY KEY (`int_artist`);
'''
cursor.execute(query_key)

cursor.execute('drop table if exists spotify_info')


tabla_spotify_info='''
            create table spotify_info(
                id varchar(50),
                name varchar(50),
                popularity int,
                type varchar(10),
                uri varchar(50),
                followers_total varchar(50)
            );
'''


cursor.execute(tabla_spotify_info)

for i in range(len(artists_idSpotify)):
    insert_query2='insert into {} ({}) values {};'\
                 .format('spotify_info', 
                         ','.join(artists_idSpotify.columns), 
                         tuple(artists_idSpotify.iloc[i].values))
    
    cursor.execute(insert_query2)
crea_db.commit()

query_key_2 = '''
ALTER TABLE `Spotify`.`spotify_info` 
ADD PRIMARY KEY (`name`);
'''
cursor.execute(query_key_2)

To collect all data of dataset, we do left join to all tables by artist name:

all_data ='''
select *
from rank_streamed_artists as ra
left join spotify_info as sp
on sp.name=ra.artist
left join past_concerts as pa
on sp.name=pa.artist
left join upcoming_concerts as up
on sp.name=up.artist
'''
cursor.execute(all_data)

For relationships= 
'''-- MySQL Workbench Synchronization
-- Generated: 2022-02-08 09:08
-- Model: New Model
-- Version: 1.0
-- Project: Name of the project
-- Author: Palmira

SET @OLD_UNIQUE_CHECKS=@@UNIQUE_CHECKS, UNIQUE_CHECKS=0;
SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS=0;
SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE='STRICT_TRANS_TABLES,NO_ZERO_IN_DATE,NO_ZERO_DATE,ERROR_FOR_DIVISION_BY_ZERO,NO_ENGINE_SUBSTITUTION';

ALTER TABLE `Spotify`.`rank_streamed_artists` 
DROP FOREIGN KEY `fk_rank_streamed_artists_past_concerts`,
DROP FOREIGN KEY `fk_rank_streamed_artists_spotify_info1`;

CREATE TABLE IF NOT EXISTS `Spotify`.`past_concerts_has_upcoming_concerts` (
  `past_concerts_int_artist` INT(11) NOT NULL,
  `upcoming_concerts_int_artist` INT(11) NOT NULL,
  PRIMARY KEY (`past_concerts_int_artist`, `upcoming_concerts_int_artist`),
  INDEX `fk_past_concerts_has_upcoming_concerts_upcoming_concerts1_idx` (`upcoming_concerts_int_artist` ASC) VISIBLE,
  INDEX `fk_past_concerts_has_upcoming_concerts_past_concerts1_idx` (`past_concerts_int_artist` ASC) VISIBLE,
  CONSTRAINT `fk_past_concerts_has_upcoming_concerts_past_concerts1`
    FOREIGN KEY (`past_concerts_int_artist`)
    REFERENCES `Spotify`.`past_concerts` (`int_artist`)
    ON DELETE NO ACTION
    ON UPDATE NO ACTION,
  CONSTRAINT `fk_past_concerts_has_upcoming_concerts_upcoming_concerts1`
    FOREIGN KEY (`upcoming_concerts_int_artist`)
    REFERENCES `Spotify`.`upcoming_concerts` (`int_artist`)
    ON DELETE NO ACTION
    ON UPDATE NO ACTION)
ENGINE = InnoDB
DEFAULT CHARACTER SET = utf8mb4
COLLATE = utf8mb4_0900_ai_ci;

ALTER TABLE `Spotify`.`rank_streamed_artists` 
DROP FOREIGN KEY `fk_rank_streamed_artists_upcoming_concerts1`;

ALTER TABLE `Spotify`.`rank_streamed_artists` ADD CONSTRAINT `fk_rank_streamed_artists_past_concerts`
  FOREIGN KEY (`past_concerts_int_artist`)
  REFERENCES `Spotify`.`past_concerts` (`int_artist`)
  ON DELETE NO ACTION
  ON UPDATE NO ACTION,
ADD CONSTRAINT `fk_rank_streamed_artists_upcoming_concerts1`
  FOREIGN KEY (`upcoming_concerts_int_artist`)
  REFERENCES `Spotify`.`upcoming_concerts` (`int_artist`)
  ON DELETE NO ACTION
  ON UPDATE NO ACTION,
ADD CONSTRAINT `fk_rank_streamed_artists_spotify_info1`
  FOREIGN KEY (`spotify_info_name`)
  REFERENCES `Spotify`.`spotify_info` (`name`)
  ON DELETE NO ACTION
  ON UPDATE NO ACTION;


SET SQL_MODE=@OLD_SQL_MODE;
SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS;
SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS;
'''

## AQUI TERMINAN LAS QUERIES

For relationships, look at queries done in mysql files, which it is summarize as follows:

ALTER TABLE `Spotify`.`rank_streamed_artists`

ADD COLUMN `past_concerts_int_artist` INT(11) NOT NULL AFTER `last_update`,

ADD COLUMN `upcoming_concerts_int_artist` INT(11) NOT NULL AFTER `past_concerts_int_artist`,

ADD COLUMN `spotify_info_name` VARCHAR(50) NOT NULL AFTER `upcoming_concerts_int_artist`,

DROP PRIMARY KEY,

ADD PRIMARY KEY (`streaming_rank`, `past_concerts_int_artist`, `upcoming_concerts_int_artist`, `spotify_info_name`),

ADD INDEX `fk_rank_streamed_artists_past_concerts_idx` (`past_concerts_int_artist` ASC) VISIBLE,

ADD INDEX `fk_rank_streamed_artists_upcoming_concerts1_idx` (`upcoming_concerts_int_artist` ASC) VISIBLE,

ADD INDEX `fk_rank_streamed_artists_spotify_info1_idx` (`spotify_info_name` ASC) VISIBLE;


ALTER TABLE `Spotify`.`rank_streamed_artists` 

ADD CONSTRAINT `fk_rank_streamed_artists_past_concerts`

  FOREIGN KEY (`past_concerts_int_artist`)
 
  REFERENCES `Spotify`.`past_concerts` (`int_artist`)
  
  ON DELETE NO ACTION
  ON UPDATE NO ACTION,
ADD CONSTRAINT `fk_rank_streamed_artists_upcoming_concerts1`

  FOREIGN KEY (`upcoming_concerts_int_artist`)
  
  REFERENCES `Spotify`.`upcoming_concerts` (`int_artist`)
  
  ON DELETE NO ACTION
  ON UPDATE NO ACTION,
ADD CONSTRAINT `fk_rank_streamed_artists_spotify_info1`

  FOREIGN KEY (`spotify_info_name`)
  
  REFERENCES `Spotify`.`spotify_info` (`name`)
  
  ON DELETE NO ACTION
  ON UPDATE NO ACTION
